move `klut_to_graph` out of `node_resynthesis/`
